F90 parallel wr2 test: collective writes
Previously, the `nf90_put_var` writes were done independently, but this
is not possible in the current NetCDF4 library, since unlimited
dimensions require collective writes.  As a result, this test was
failing on our system.

We have resolved the problem by switing from `nf90_independent` to
`nf90_collective` writes, and removing the rank 0 write test, which was
blocking the (now collective) `nf90_put_var` call.