Get filter test to work with Fortran
re: https://github.com/Unidata/netcdf-fortran/issues/169

This required a change both to netcdf-c and to netcdf-fortran.
For netcdf-c, see PR https://github.com/Unidata/netcdf-c/pull/1387

The fortran fix required creating a shell script: nf_test4/run_tst_filter.sh.
This script ensures that HDF5_PLUGIN_DIR is set to nf-config --prefix.
This assumes that you are using the version of netcdf-c where the bzip2
plugin is installed in the --prefix/lib dir as part of the netcdf-c
library install. This may not work correctly for Windows or OSX.